+++ ktx/__init__.py
@@ -1,6 +1,8 @@
 from .vars import get_current_ctx, get_current_ctx_or_none
+from .wrap import ctx_wrap
 
 __all__ = [
     "get_current_ctx",
     "get_current_ctx_or_none",
+    "ctx_wrap",
 ]